modeling/ns_model/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
docopt==0.6.2
etaprogress==1.1.1
filelock==3.9.0
gitdb==4.0.10
GitPython==3.1.30
gtimer==1.0.0b5
idna==3.4
jsonpickle==1.5.2
lmdb==1.4.0
munch==2.5.0
networkx==3.0
numpy==1.24.1
opencv-python==4.7.0.68
packaging==23.0
pandas==1.5.2
Pillow==9.4.0
py-cpuinfo==9.0.0
python-dateutil==2.8.2
pytz==2022.7
requests==2.28.2
revtok==0.0.3
sacred==0.8.2
scikit-video==1.1.11
scipy==1.10.0
six==1.16.0
smmap==5.0.0
termcolor==2.2.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
vocab==0.0.5
wrapt==1.14.1
dotmap==1.3.30
boto3